% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{linal_query}
\alias{linal_query}
\title{Run a DSL command expected to return a table and return a `data.frame`}
\usage{
linal_query(conn, dsl)
}
\arguments{
\item{conn}{A `linal_connection` from `linal_connect()`.}

\item{dsl}{The DSL command string.}
}
\value{
A `data.frame`. Raises a `linal_error` if the result isn't
  table-shaped.
}
\description{
Run a DSL command expected to return a table and return a `data.frame`
}
