% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linaldb.server-package.R
\docType{package}
\name{linaldb.server-package}
\alias{linaldb.server}
\alias{linaldb.server-package}
\title{linaldb.server: R Client for LINALDB}
\description{
Talks to a running `linal serve` instance over its HTTP API. See
`PYTHON_R_INTEROP_PLAN.md` (checkpoints 3-4) in the parent repository
for the real client implementation and `clients/CONTRACT.md` for the
wire contract this package will implement against.
}
\details{
Scaffolding only as of this file -- no functions defined yet.
}
\author{
\strong{Maintainer}: gorigami \email{develop@gorigami.xyz}

Authors:
\itemize{
  \item gorigami \email{develop@gorigami.xyz}
}

}
\keyword{internal}
